+++ lib/interceptors/network/network_log.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 class NetworkLog {
   final String method;
   final String url;
@@ -18,4 +20,27 @@ class NetworkLog {
     required this.duration,
     this.isError = false,
   });
+
+  Map<String, dynamic> toJson() {
+    return {
+      'method': method,
+      'url': url,
+      'requestHeaders': requestHeaders,
+      'requestBody': requestBody,
+      'statusCode': statusCode,
+      'responseBody': _parseIfJson(responseBody),
+      'duration': duration.inMilliseconds,
+      'isError': isError,
+    };
+  }
+
+  dynamic _parseIfJson(String? input) {
+    if (input == null) return null;
+
+    try {
+      return json.decode(input);
+    } catch (_) {
+      return input; // return as-is if not valid JSON
+    }
+  }
 }