+++ lib/interceptors/performance/cold_start_timer.dart
@@ -1,3 +1,4 @@
+import '../../core/logger_interface.dart';
 import '../../flutter_dev_toolkit.dart';
 
 class ColdStartTimer {